# add max_demand param