add max_demands parameter to Flow.from_enumerable

**Benchmark**

Without add max_demands: 100

```
Name                 ips        average  deviation         median         99th %
find_typos        0.0285        35.06 s     ±0.00%        35.06 s        35.06 s
```

With add max_demands: 100

```
Name                 ips        average  deviation         median         99th %
find_typos        0.0488        20.48 s     ±0.00%        20.48 s        20.4
```

Time performance: ~ -42.%