@samuelpordeus 2019-12-12
@g13ydson Thanks for the PR! that's an awesome improvement. What about memory consumption?