@msramos 2019-12-12
Memory consumption shouldn't be affected. The demand param just tells flow to increase the batch size between stages.